Rename component -> state
