--- runtime-light/component/component.cmake
@@ -1,1 +1,0 @@
-prepend(RUNTIME_LIGHT_COMPONENT_SRC component/ component.cpp init-functions.cpp)
