+++ runtime-light/core/kphp-core-impl/kphp-core-context.cpp
@@ -3,7 +3,7 @@
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-light/component/component.h"
+#include "runtime-light/state/instance-state.h"
 
 constexpr string_size_type initial_minimum_string_buffer_length = 1024;
 constexpr string_size_type initial_maximum_string_buffer_length = (1 << 24);
