+++ runtime-light/scheduler/scheduler.cpp
@@ -11,7 +11,7 @@
 #include <utility>
 #include <variant>
 
-#include "runtime-light/component/component.h"
+#include "runtime-light/state/instance-state.h"
 
 // === SimpleCoroutineScheduler ===================================================================
