+++ runtime-light/state/image-state.h
