+++ runtime-light/state/state.cmake
@@ -1,0 +1,1 @@
+prepend(RUNTIME_LIGHT_STATE_SRC state/ instance-state.cpp init-functions.cpp)
