+++ runtime-light/stdlib/component/component-api.cpp
@@ -8,10 +8,10 @@
 
 #include "runtime-common/core/runtime-core.h"
 #include "runtime-common/core/utils/kphp-assert-core.h"
-#include "runtime-light/component/component.h"
 #include "runtime-light/coroutine/awaitable.h"
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/k2-platform/k2-api.h"
+#include "runtime-light/state/instance-state.h"
 #include "runtime-light/streams/streams.h"
 
 // === component query client interface ===========================================================
