+++ runtime-light/stdlib/component/component-api.h
@@ -10,8 +10,8 @@
 
 #include "runtime-common/core/class-instance/refcountable-php-classes.h"
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-light/component/component.h"
 #include "runtime-light/coroutine/task.h"
+#include "runtime-light/state/instance-state.h"
 
 // === ComponentQuery =============================================================================
