+++ runtime-light/stdlib/curl/curl-context.cpp
@@ -4,7 +4,7 @@
 
 #include "runtime-light/stdlib/curl/curl-context.h"
 
-#include "runtime-light/component/component.h"
+#include "runtime-light/state/instance-state.h"
 
 CurlInstanceState &CurlInstanceState::get() noexcept {
   return InstanceState::get().curl_instance_state;
