+++ runtime-light/stdlib/exit/exit-functions.cpp
@@ -6,8 +6,8 @@
 
 #include <cstdint>
 
-#include "runtime-light/component/component.h"
 #include "runtime-light/k2-platform/k2-api.h"
+#include "runtime-light/state/instance-state.h"
 
 task_t<void> f$exit(const mixed &v) noexcept { // TODO: make it synchronous
   auto &instance_st{InstanceState::get()};
