+++ runtime-light/stdlib/fork/fork-context.cpp
@@ -4,7 +4,7 @@
 
 #include "runtime-light/stdlib/fork/fork-context.h"
 
-#include "runtime-light/component/component.h"
+#include "runtime-light/state/instance-state.h"
 
 ForkInstanceState &ForkInstanceState::get() noexcept {
   return InstanceState::get().fork_instance_state;
