+++ runtime-light/stdlib/job-worker/job-worker-api.h
@@ -7,8 +7,8 @@
 #include <cstdint>
 
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-light/component/component.h"
 #include "runtime-light/coroutine/task.h"
+#include "runtime-light/state/instance-state.h"
 
 // === Client =====================================================================================
