+++ runtime-light/stdlib/job-worker/job-worker-client-context.cpp
@@ -4,7 +4,7 @@
 
 #include "runtime-light/stdlib/job-worker/job-worker-client-context.h"
 
-#include "runtime-light/component/component.h"
+#include "runtime-light/state/instance-state.h"
 
 JobWorkerClientInstanceState &JobWorkerClientInstanceState::get() noexcept {
   return InstanceState::get().job_worker_client_instance_state;
